Fix FZF_TAB_HOME when a symlink is sourced
FZF_TAB_HOME will contain a wrong directory when instead of fzf-tab.zsh
or fzf-tab.plugin.zsh, a symlink pointing to one of the two is sourced
and the symlink lays in a different directory.
In this case it will not contain the repository root but the directory
in which the symlink is.